pyproject.toml
[tool.poetry]
name = "inso-bootstrap-cli"
# keep manually in sync with __init__.py > __version__
version = "2.4.0"
description = "A CLI to deploy a CDF Project to bootstrap CDF Groups scoped with Data Sets and RAW DBs"
authors = ["Peter Arwanitis <peter.arwanitis@cognite.com>", "Tugce Ozgur Oztetik <tugce.oztetik@cognite.com>"]
license = "Apache-2.0"
readme = "README.md"
repository = "https://github.com/cognitedata/inso-bootstrap-cli"

packages = [
    { include="incubator", from="." },
]

[tool.black]
line-length = 120
target_version = ['py39']
include = '\.py$'

[tool.isort]
# https://pycqa.github.io/isort/docs/configuration/black_compatibility.html
profile = "black"
line_length=120                # corresponds to -w  flag
multi_line_output=3            # corresponds to -m  flag
include_trailing_comma=true    # corresponds to -tc flag
skip_glob = '^((?!py$).)*$'    # this makes sort all Python files
known_third_party = ["click", "pytest"]

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^0.21.1"
pydantic = "^1.10.4"
dependency-injector = {version = "^4.41.0", extras = ["yaml"]}
click = "^8.1.3"
# TODO: 230125 required?
# types-retry = "^0.1.5"
cognite-sdk = {version = "^5.2.0", extras = ["pandas"]}
# TODO: 230125 was this required for windows build?!
setuptools = "^62.1.0"
rich = "^13.2.0"

[tool.poetry.dev-dependencies]
black = "^22.12.0"
isort = "^5.11.4"
flake8 = "^6.0.0"
mypy = "^0.991"
# autoflake = "=1.3"
pre-commit = "^3.0.0"
pytest = "^7.2.1"
# pytest-cov = "^3.0.0"

[tool.poetry.scripts]
bootstrap-cli = "incubator.bootstrap_cli.__main__:main"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.semantic_release]
version_toml = [
"pyproject.toml:tool.poetry.version"
]
